+++ registry/handlers/tags.go
@@ -1,6 +1,7 @@
 package handlers
 
 import (
+	"encoding/json"
 	"net/http"
 
 	"github.com/docker/distribution"
@@ -18,7 +19,7 @@ func tagsDispatcher(ctx *Context, r *http.Request) http.Handler {
 	}
 
 	thandler := handlers.MethodHandler{
-		"GET": http.HandlerFunc(tagsHandler.GetTags),
+		"GET": http.HandlerFunc(tagsHandler.GetTagsCompat),
 	}
 
 	if !ctx.readOnly {
@@ -33,7 +34,28 @@ type tagsHandler struct {
 	Tag string
 }
 
-func (th *tagsHandler) GetTags(w http.ResponseWriter, r *http.Request) {
+// tagsDispatcher constructs the tags handler api endpoint.
+func tagsListDispatcher(ctx *Context, r *http.Request) http.Handler {
+	tagsListHandler := &tagsListHandler{
+		Context: ctx,
+	}
+
+	return handlers.MethodHandler{
+		"GET": http.HandlerFunc(tagsListHandler.GetTags),
+	}
+}
+
+// tagsHandler handles requests for lists of tags under a repository name.
+type tagsListHandler struct {
+	*Context
+}
+
+type tagsListAPIResponse struct {
+	Name string   `json:"name"`
+	Tags []string `json:"tags"`
+}
+
+func (th *tagsHandler) GetTagsCompat(w http.ResponseWriter, r *http.Request) {
 	ctxu.GetLogger(th).Debug("GetTags")
 
 	if th.Tag == "list" {
@@ -73,4 +95,34 @@ func (th *tagsHandler) DeleteTag(w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.WriteHeader(http.StatusAccepted)
+}
+
+// GetTags returns a json list of tags for a specific image name.
+func (th *tagsListHandler) GetTags(w http.ResponseWriter, r *http.Request) {
+	defer r.Body.Close()
+
+	tagService := th.Repository.Tags(th)
+	tags, err := tagService.All(th)
+	if err != nil {
+		switch err := err.(type) {
+		case distribution.ErrRepositoryUnknown:
+			th.Errors = append(th.Errors, v2.ErrorCodeNameUnknown.WithDetail(map[string]string{"name": th.Repository.Named().Name()}))
+		case errcode.Error:
+			th.Errors = append(th.Errors, err)
+		default:
+			th.Errors = append(th.Errors, errcode.ErrorCodeUnknown.WithDetail(err))
+		}
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+
+	enc := json.NewEncoder(w)
+	if err := enc.Encode(tagsListAPIResponse{
+		Name: th.Repository.Named().Name(),
+		Tags: tags,
+	}); err != nil {
+		th.Errors = append(th.Errors, errcode.ErrorCodeUnknown.WithDetail(err))
+		return
+	}
 }
